requirements.txt
anyio==3.4.0
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
bcrypt==3.2.0
boto3==1.20.24
botocore==1.23.24
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
chromedriver-binary==97.0.4692.36.0
chromedriver-binary-auto==0.1.1
click==8.0.3
cryptography==36.0.1
ecdsa==0.17.0
fastapi==0.70.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
idna==3.3
jmespath==0.10.0
outcome==1.1.0
passlib==1.7.4
psycopg2==2.9.2
psycopg2-binary==2.9.2
pyasn1==0.4.8
pybites-carbon==0.4
pybites-tools==0.1
pycparser==2.21
pydantic==1.8.2
pyOpenSSL==21.0.0
pyperclip==1.8.2
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.19.2
python-jose==3.3.0
requests==2.26.0
rsa==4.8
s3transfer==0.5.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.28
sqlalchemy2-stubs==0.0.2a19
sqlmodel==0.0.5
starlette==0.16.0
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.0.1
urllib3==1.26.7
uvicorn==0.16.0
wsproto==1.0.0
